Make AWS SDK framework build with OS X 64-bit target.

+++ src/include/S3/S3ListPartsRequest.h
@@ -22,8 +22,8 @@
  */
 @interface S3ListPartsRequest:S3Request {
     NSString  *uploadId;
-    NSInteger maxParts;
-    NSInteger partNumberMarker;
+    int maxParts;
+    int partNumberMarker;
 }
 
 /** The uploadId of the multipart upload to list */
